Make Get a strict state monad, using unboxed tuples for speed
Contrary to comments in the code, the bind operator for Get was lazy.
Making it strict and using unboxed tuples for the return type improves
performance for me.

One possible concern is whether the lazy Get monad is being relied on
to avoid stack overflow, but looking at the fix that was needed
recently for replicateM I'm presuming that this is not an issue.