Make Buffer type smaller and support faster indexing
The new Buffer type has only two fields, causing less allocation and a
smaller number of parameters being passed between functions when the
Buffer is unboxed.

The new buffer type supports more efficient indexing as it avoids
having to add an offset on each access.  This is made possible by
moving to the call to withForeignPtr up to a driver loop.

Benchmark results on a 64-bit machine:

Before
======

Binary (de)serialisation benchmarks:
100MB of Word8  in chunks of 16 (Host endian):  153.8 MB/s write
100MB of Word16 in chunks of 16 (Host endian):  312.5 MB/s write
100MB of Word32 in chunks of 16 (Host endian):  500.0 MB/s write
100MB of Word64 in chunks of 16 (Host endian): 1000.0 MB/s write

After
=====

Binary (de)serialisation benchmarks:
100MB of Word8  in chunks of 16 (Host endian):  192.3 MB/s write
100MB of Word16 in chunks of 16 (Host endian):  400.0 MB/s write
100MB of Word32 in chunks of 16 (Host endian):  588.2 MB/s write
100MB of Word64 in chunks of 16 (Host endian): 1428.6 MB/s write

This also improves write performance for the [Word8] benchmark with
15% and will make further unboxing optimizations possible in the
future.

As an added benefit buffer allocation is now lifted up to the top,
making it possible to let users control it (although that
functionality is not included in this patch.)