Implement lookAhead.
Issue #17. Test present, documentation still missing.
This function was removed with the 0.6.* branch, and had quite few users.
The plan was to use <|> instead, but lookAhead was missed since it
offers some functionality that <|> cannot provide without larger restructuring of the application.
lookAhead can be implemented in binary without affecting any other functionality,
so here it is again.

This also makes 0.6.* more backwards compatible with earlier versions,
this saving the developer from #if MIN_VERSION hacks.