Bug fix isolate; keep labels from within failing isolate
When an isolated decoder failed due to over consuming input, any
labels set within that decoder were lost since demanding
more input would fail within the isolate-decoder runner and not in
the isolated decoder itself.
With this change the signal of lack of more input is passed
into the isolated decoder which in turn will fail, keeping any
labels up to that point.