Fail rather than throw exception when decoding Bool and Ordering
Decoding Bool and Ordering was not graceful in case of invalid input.
Trying to decode (2 :: Word8) to a Bool would throw an exception
in GHC.Enum.toEnum as only the values 0 and 1 are defined.

We work around that in this patch by not using toEnum. In case
of unexpected input, we fail using 'fail'.

This fixes #108.