Use larger type if size doesn't fit.
The current implementation uses the output type to store the number
of constructors (e.g. parameter size of putSum). This will fail if
we try to serialize a type with 256 constructors, which can be encoded
in Word8 but its size does not. This is slightly inefficient: We could
use Word8 if getSum and putSum were implemented differently (which would
make them more complicated and probably slower).